Refactor to allow iOS to import the ai_chat components Add empty autocomplete provider implementation for Leo
