+++ browser/ui/webui/ai_chat/ai_chat_ui.cc
@@ -10,8 +10,8 @@
 #include "brave/browser/ui/side_panel/ai_chat/ai_chat_side_panel_utils.h"
 #include "brave/browser/ui/webui/ai_chat/ai_chat_ui_page_handler.h"
 #include "brave/browser/ui/webui/brave_webui_source.h"
-#include "brave/components/ai_chat/browser/constants.h"
 #include "brave/components/ai_chat/common/pref_names.h"
+#include "brave/components/ai_chat/core/constants.h"
 #include "brave/components/ai_chat/resources/page/grit/ai_chat_ui_generated_map.h"
 #include "brave/components/constants/webui_url_constants.h"
 #include "brave/components/l10n/common/localization_util.h"
