+++ browser/ui/webui/ai_chat/ai_chat_ui_page_handler.cc
@@ -12,11 +12,11 @@
 #include "base/notreached.h"
 #include "base/strings/utf_string_conversions.h"
 #include "brave/components/ai_chat/browser/ai_chat_tab_helper.h"
-#include "brave/components/ai_chat/browser/constants.h"
-#include "brave/components/ai_chat/browser/models.h"
 #include "brave/components/ai_chat/common/mojom/ai_chat.mojom-shared.h"
 #include "brave/components/ai_chat/common/mojom/ai_chat.mojom.h"
 #include "brave/components/ai_chat/common/pref_names.h"
+#include "brave/components/ai_chat/core/constants.h"
+#include "brave/components/ai_chat/core/models.h"
 #include "chrome/browser/favicon/favicon_service_factory.h"
 #include "chrome/browser/profiles/profile.h"
 #include "components/favicon/core/favicon_service.h"
