+++ components/ai_chat/browser/DEPS
@@ -3,6 +3,5 @@ include_rules = [
   "+services/data_decoder/public",
   "+services/network/public",
   "+services/service_manager/public",
-  "+absl/types/optional.h",
   "+ui",
 ]
