+++ components/ai_chat/browser/ai_chat_tab_helper.cc
@@ -17,16 +17,16 @@
 #include "base/notreached.h"
 #include "base/ranges/algorithm.h"
 #include "base/strings/string_util.h"
-#include "brave/components/ai_chat/browser/ai_chat_metrics.h"
-#include "brave/components/ai_chat/browser/engine/engine_consumer.h"
-#include "brave/components/ai_chat/browser/engine/engine_consumer_claude.h"
-#include "brave/components/ai_chat/browser/engine/engine_consumer_llama.h"
-#include "brave/components/ai_chat/browser/models.h"
 #include "brave/components/ai_chat/browser/page_content_fetcher.h"
 #include "brave/components/ai_chat/common/features.h"
 #include "brave/components/ai_chat/common/mojom/ai_chat.mojom-shared.h"
 #include "brave/components/ai_chat/common/mojom/ai_chat.mojom.h"
 #include "brave/components/ai_chat/common/pref_names.h"
+#include "brave/components/ai_chat/core/ai_chat_metrics.h"
+#include "brave/components/ai_chat/core/engine/engine_consumer.h"
+#include "brave/components/ai_chat/core/engine/engine_consumer_claude.h"
+#include "brave/components/ai_chat/core/engine/engine_consumer_llama.h"
+#include "brave/components/ai_chat/core/models.h"
 #include "components/favicon/content/content_favicon_driver.h"
 #include "components/grit/brave_components_strings.h"
 #include "components/prefs/pref_service.h"
