+++ components/ai_chat/common/mojom/ai_chat.mojom
@@ -33,7 +33,7 @@ enum ModelEngineType {
 };
 
 // When adding or modifying category names, associated UI strings must also
-// be provided via ai_chat_ui_strings.grdp and ai_chat/browser/constants.cc
+// be provided via ai_chat_ui_strings.grdp and ai_chat/core/constants.cc
 enum ModelCategory {
   CHAT,
 };
