+++ components/ai_chat/core/BUILD.gn
@@ -11,12 +11,21 @@ if (is_official_build) {
   assert(brave_ai_chat_endpoint != "")
 }
 
-static_library("browser") {
+static_library("core") {
   sources = [
-    "ai_chat_tab_helper.cc",
-    "ai_chat_tab_helper.h",
-    "page_content_fetcher.cc",
-    "page_content_fetcher.h",
+    "ai_chat_metrics.cc",
+    "ai_chat_metrics.h",
+    "constants.cc",
+    "constants.h",
+    "engine/engine_consumer.h",
+    "engine/engine_consumer_claude.cc",
+    "engine/engine_consumer_claude.h",
+    "engine/engine_consumer_llama.cc",
+    "engine/engine_consumer_llama.h",
+    "engine/remote_completion_client.cc",
+    "engine/remote_completion_client.h",
+    "models.cc",
+    "models.h",
   ]
 
   deps = [
@@ -24,19 +33,18 @@ static_library("browser") {
     "//brave/components/ai_chat/common",
     "//brave/components/ai_chat/common/buildflags",
     "//brave/components/ai_chat/common/mojom",
-    "//brave/components/ai_chat/core",
     "//brave/components/api_request_helper",
     "//brave/components/constants",
+    "//brave/components/p3a_utils",
     "//brave/components/resources:strings_grit",
-    "//components/favicon/content",
-    "//components/favicon/core",
+    "//brave/components/time_period_storage",
     "//components/prefs",
     "//components/user_prefs",
-    "//content/public/browser",
     "//net/traffic_annotation",
     "//services/data_decoder/public/cpp",
     "//services/network/public/cpp",
     "//services/service_manager/public/cpp",
+    "//third_party/abseil-cpp:absl",
     "//ui/base",
     "//url",
   ]
