+++ components/ai_chat/core/DEPS
@@ -1,0 +1,4 @@
+include_rules = [
+  "+services/network/public",
+  "+absl/types/optional.h",
+]
