+++ components/ai_chat/core/ai_chat_metrics_unittest.cc
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#include "brave/components/ai_chat/browser/ai_chat_metrics.h"
+#include "brave/components/ai_chat/core/ai_chat_metrics.h"
 
 #include <memory>
