+++ components/ai_chat/core/constants.cc
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#include "brave/components/ai_chat/browser/constants.h"
+#include "brave/components/ai_chat/core/constants.h"
 #include "base/strings/strcat.h"
 
 namespace ai_chat {
