+++ components/ai_chat/core/constants.h
@@ -3,8 +3,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-#ifndef BRAVE_COMPONENTS_AI_CHAT_BROWSER_CONSTANTS_H_
-#define BRAVE_COMPONENTS_AI_CHAT_BROWSER_CONSTANTS_H_
+#ifndef BRAVE_COMPONENTS_AI_CHAT_CORE_CONSTANTS_H_
+#define BRAVE_COMPONENTS_AI_CHAT_CORE_CONSTANTS_H_
 
 #include "components/grit/brave_components_strings.h"
 #include "ui/base/webui/web_ui_util.h"
@@ -15,4 +15,4 @@ base::span<const webui::LocalizedString> GetLocalizedStrings();
 
 }  // namespace ai_chat
 
-#endif  // BRAVE_COMPONENTS_AI_CHAT_BROWSER_CONSTANTS_H_
+#endif  // BRAVE_COMPONENTS_AI_CHAT_CORE_CONSTANTS_H_
