+++ components/ai_chat/core/engine/engine_consumer.h
@@ -3,8 +3,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#ifndef BRAVE_COMPONENTS_AI_CHAT_BROWSER_ENGINE_ENGINE_CONSUMER_H_
-#define BRAVE_COMPONENTS_AI_CHAT_BROWSER_ENGINE_ENGINE_CONSUMER_H_
+#ifndef BRAVE_COMPONENTS_AI_CHAT_CORE_ENGINE_ENGINE_CONSUMER_H_
+#define BRAVE_COMPONENTS_AI_CHAT_CORE_ENGINE_ENGINE_CONSUMER_H_
 
 #include <string>
 #include <vector>
@@ -60,4 +60,4 @@ class EngineConsumer {
 
 }  // namespace ai_chat
 
-#endif  // BRAVE_COMPONENTS_AI_CHAT_BROWSER_ENGINE_ENGINE_CONSUMER_H_
+#endif  // BRAVE_COMPONENTS_AI_CHAT_CORE_ENGINE_ENGINE_CONSUMER_H_
